remove old resource tree
